remove old p5js files
